+++ lib/src/4337/providers.dart
@@ -1,28 +1,12 @@
 part of '../../variance.dart';
 
 class BundlerProvider implements BundlerProviderBase {
-  /// Set of Ethereum RPC methods supported by the SmartWallet SDK.
-  static final Set<String> methods = {
-    'eth_chainId',
-    'eth_sendUserOperation',
-    'eth_estimateUserOperationGas',
-    'eth_getUserOperationByHash',
-    'eth_getUserOperationReceipt',
-    'eth_supportedEntryPoints',
-  };
-
-  final int _chainId;
-  final String _bundlerUrl;
-  final RPCProviderBase _bundlerRpc;
+  final Chain _chain;
+  final RPCProviderBase _rpc;
 
   late final bool _initialized;
 
-  Entrypoint? entrypoint;
-
-  BundlerProvider(Chain chain, RPCProviderBase bundlerRpc)
-      : _chainId = chain.chainId,
-        _bundlerUrl = chain.bundlerUrl!,
-        _bundlerRpc = bundlerRpc {
+  BundlerProvider(this._chain, this._rpc) {
     _initializeBundlerProvider();
   }
 
@@ -30,7 +14,7 @@ class BundlerProvider implements BundlerProviderBase {
   Future<UserOperationGas> estimateUserOperationGas(
       Map<String, dynamic> userOp, EthereumAddress entrypoint) async {
     require(_initialized, "estimateUserOpGas: Wallet Provider not initialized");
-    final opGas = await _bundlerRpc.send<Map<String, dynamic>>(
+    final opGas = await _rpc.send<Map<String, dynamic>>(
         'eth_estimateUserOperationGas', [userOp, entrypoint.hex]);
     return UserOperationGas.fromMap(opGas);
   }
@@ -38,7 +22,7 @@ class BundlerProvider implements BundlerProviderBase {
   @override
   Future<UserOperationByHash> getUserOperationByHash(String userOpHash) async {
     require(_initialized, "getUserOpByHash: Wallet Provider not initialized");
-    final opExtended = await _bundlerRpc
+    final opExtended = await _rpc
         .send<Map<String, dynamic>>('eth_getUserOperationByHash', [userOpHash]);
     return UserOperationByHash.fromMap(opExtended);
   }
@@ -46,87 +30,35 @@ class BundlerProvider implements BundlerProviderBase {
   @override
   Future<UserOperationReceipt> getUserOpReceipt(String userOpHash) async {
     require(_initialized, "getUserOpReceipt: Wallet Provider not initialized");
-    final opReceipt = await _bundlerRpc.send<Map<String, dynamic>>(
+    final opReceipt = await _rpc.send<Map<String, dynamic>>(
         'eth_getUserOperationReceipt', [userOpHash]);
     return UserOperationReceipt.fromMap(opReceipt);
-  }
-
-  @override
-  void initializeWithEntrypoint(Entrypoint ep) {
-    entrypoint = ep;
   }
 
   @override
   Future<UserOperationResponse> sendUserOperation(
       Map<String, dynamic> userOp, EthereumAddress entrypoint) async {
     require(_initialized, "sendUserOp: Wallet Provider not initialized");
-    final opHash = await _bundlerRpc
+    final opHash = await _rpc
         .send<String>('eth_sendUserOperation', [userOp, entrypoint.hex]);
-    return UserOperationResponse(opHash, wait);
+    return UserOperationResponse(opHash);
   }
 
   @override
   Future<List<String>> supportedEntryPoints() async {
     final entrypointList =
-        await _bundlerRpc.send<List<dynamic>>('eth_supportedEntryPoints');
+        await _rpc.send<List<dynamic>>('eth_supportedEntryPoints');
     return List.castFrom(entrypointList);
   }
 
-  @override
-  Future<FilterEvent?> wait({int millisecond = 0}) async {
-    if (millisecond == 0) {
-      return null;
-    }
-    require(entrypoint != null, "Entrypoint required! use Wallet.init");
-    final block = await entrypoint!.client.getBlockNumber();
-    final end = DateTime.now().millisecondsSinceEpoch + millisecond;
-
-    return await Isolate.run(() async {
-      while (DateTime.now().millisecondsSinceEpoch < end) {
-        final filterEvent = await entrypoint!.client
-            .events(
-              FilterOptions.events(
-                contract: entrypoint!.self,
-                event: entrypoint!.self.event('UserOperationEvent'),
-                fromBlock: BlockNum.exact(block - 100),
-              ),
-            )
-            .take(1)
-            .first;
-        if (filterEvent.transactionHash != null) {
-          return filterEvent;
-        }
-        await Future.delayed(Duration(milliseconds: millisecond));
-      }
-      return null;
-    });
-  }
-
   Future _initializeBundlerProvider() async {
-    final chainId = await _bundlerRpc
+    final chainId = await _rpc
         .send<String>('eth_chainId')
         .then(BigInt.parse)
         .then((value) => value.toInt());
-    require(chainId == _chainId,
-        "bundler $_bundlerUrl is on chainId $chainId, but provider is on chainId $_chainId");
+    require(chainId == _chain.chainId,
+        "bundler ${_rpc.url} is on chainId $chainId, but provider is on chainId ${_chain.chainId}");
     _initialized = true;
-  }
-
-  /// Validates if the provided method is a supported RPC method.
-  ///
-  /// Parameters:
-  ///   - `method`: The Ethereum RPC method to validate.
-  ///
-  /// Throws:
-  ///   - A [Exception] if the method is not a valid supported method.
-  ///
-  /// Example:
-  /// ```dart
-  /// validateBundlerMethod('eth_sendUserOperation');
-  /// ```
-  static validateBundlerMethod(String method) {
-    require(methods.contains(method),
-        "validateMethod: method ::'$method':: is not a valid method");
   }
 }